tests/test_config.py
"""Test configuration management."""

import tempfile
from pathlib import Path

import pytest
import yaml

# Note: These tests will fail until dependencies are installed
# They are provided as examples of the testing structure


def test_config_creation():
    """Test creating a default configuration."""
    # This is a placeholder test that doesn't require imports
    assert True


def test_config_loading():
    """Test loading configuration from YAML."""
    # Create a temporary YAML file
    with tempfile.NamedTemporaryFile(mode='w', suffix='.yaml', delete=False) as f:
        config_data = {
            "threads": 4,
            "input": {
                "sra_list": "test.txt",
                "fastq_dir": None,
                "paired_end": False,
                "metadata": None
            },
            "output": {
                "output_dir": "/tmp/test_output",
                "log_level": "INFO"
            }
        }
        yaml.dump(config_data, f)
        temp_path = Path(f.name)
    
    try:
        # Test would load config here
        # config = load_config(temp_path)
        # assert config.threads == 4
        assert temp_path.exists()
    finally:
        temp_path.unlink()


def test_pipeline_step_creation():
    """Test creating pipeline steps."""
    # Placeholder test
    assert True


if __name__ == "__main__":
    pytest.main([__file__])
